Fixed failing unit test.
Turns out I managed to use the wrong version of Akavache when writing the service layer (not hard when there are 8 different packages and Akavache.Portable looks legit, but seems to be really old and outdated). So I swapped over from Akavache.Portable to Akavache.Core and fixed various bits and pieces up accordingly.