Greatly simplified the API and implementation.
I realised whilst refactoring that it is entirely unnecessary to depend on an event type at all. As long as the parser is able to get the actual value (formerly known as "current" value) then it has all it needs to know in order to match or not. This way the caller is in control of whether the actual value comes from a NumberedEvent or something else. The implications of this are that the methods in the parser no longer need to be generic and the usage of the NumericalConstraintParser is much simpler and more logical, as evident in the unit test changes.